blog-service/requirements.txt
Django==5.0
djangorestframework==3.15
psycopg2-binary==2.9.7
django-redis==5.4.0
django-filter==23.5
python-slugify==8.0.1
gunicorn==21.2.0
django-cors-headers==4.3.1